First checkin of the basic AutoComplete $.Plugin
It is set up as a jquery plugin that has a number of paramaters capable
of being passed to it, and those will be documented upon it's v0.1
completion.  This will be spun off after the v0.1 or v0.2 implementation
into a seperate repository as a jquery / backbone autocomplete plugin.
It is very very very minimal.  And will stay true to that.  It is also
very reusable and configurable and will stay true to that as well.