Added trigger option to plugin.
I added the ability for a user to pass in a trigger paramater, with the
trigger character.  And then the autocomplete plugin will make sure to
only start searching once that has been keyed in, and it will strip it
(obviously) before sending results back to the server.  It will also
clear the results if they were searching with a trigger and stopped to
search without a trigger (using the user passed in class option).