Added depth check.
Added the depth check with the help of @polastre, using an associative
array to add the depth property to the comment.  If that depth is
greater of or equal to the child of a topic's comment, then we don't
render the reply form or the form logic for that child-child comment.